fix(sources-postgres): decode WAL keys to match the read-back variant
A live change's key is decoded twice: from the WAL by Postgres type OID
(`typed_value`), and from a query result by SQL type name (`decode_column`)
when the row is rebuilt. The two disagreed on the `GenericValue` variant —
`uuid` decoded as `String` vs read back as `Uuid`, and every integer width
collapsed to `BigInt` vs read back as `Int`/`SmallInt`. The engine matches a
live change's key against the rebuilt row's key by variant-exact equality in
`build_many`, so the match missed and every live insert/update on such a table
was written as a tombstone — the document silently vanished. `bigint`/`text`
keys agreed, masking the bug.

Mirror `decode_column` in `typed_value`: width-exact integers, split floats,
and decode `uuid`.

Add a full-pipeline e2e (`libs/2-engine/tests/pipeline.rs`): real Postgres ->
the real `Engine::run` -> a real OpenSearch sink, reading the index back to
assert the actual document after insert/update/delete across uuid/int/bigint/
text keys, soft-delete tombstoning (set -> removed, cleared -> restored), and
backfill (active seeded, soft-deleted skipped). Reverting the decoder fix makes
it fail.

Closes #53

+++ libs/1-sources/1-postgres/src/cdc/pgoutput/tests.rs
@@ -61,7 +61,7 @@ fn insert_yields_only_key_columns() {
     let key = row_key(&rel, &new).unwrap();
     assert_eq!(key.0.len(), 1);
     assert_eq!(key.0[0].0.as_ref(), "id");
-    assert_eq!(key.0[0].1, GenericValue::BigInt(42)); // id is int4 (oid 23)
+    assert_eq!(key.0[0].1, GenericValue::Int(42)); // id is int4 (oid 23)
 }
 
 #[test]
@@ -81,7 +81,7 @@ fn delete_uses_old_key_tuple() {
         panic!("expected Delete");
     };
     let key = row_key(&rel, &old).unwrap();
-    assert_eq!(key.0[0].1, GenericValue::BigInt(42)); // id is int4 (oid 23)
+    assert_eq!(key.0[0].1, GenericValue::Int(42)); // id is int4 (oid 23)
 }
 
 #[test]
